+++ pkg/controller/backend.go
@@ -1103,6 +1103,7 @@ func createRuleAction(rl *Rule, rulesData *as3Rule) {
 
 //Extract virtual address and port from host URL
 func extractVirtualAddressAndPort(str string) (string, int) {
+
 	destination := strings.Split(str, "/")
 	// split separator is in accordance with SetVirtualAddress function - ipv4/6 format
 	ipPort := strings.Split(destination[len(destination)-1], ":")