+++ docs/RELEASE-NOTES.rst
@@ -9,7 +9,7 @@ Added Functionality
 * Added Support for AS3 3.38.0
 * CRD:
         * AllowSourceRange support for VirtualServer CRs and Policy CRs. See `Examples <https://github.com/F5Networks/k8s-bigip-ctlr/tree/master/docs/config_examples/customResource/>`_
-
+* Added support to configure netmask for Virtual Server for Ingress. See `Examples <https://github.com/F5Networks/k8s-bigip-ctlr/tree/master/docs/config_examples/ingress/>`_
 Bug Fixes
 ````````````
 * :issues:`2361` Allow monitoring of an alias port in VirtualServer and TransportServer